backend/requirements.txt
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database and ORM
supabase==2.0.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# HTTP requests and API clients
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1

# Data processing and ML
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
tensorflow==2.15.0
keras==2.15.0
joblib==1.3.2

# Time series and forecasting
statsmodels==0.14.0
plotly==5.17.0

# Notification services
twilio==8.10.3
firebase-admin==6.2.0

# Satellite and geospatial data
sentinelhub==3.9.1
rasterio==1.3.9
geopandas==0.14.1
shapely==2.0.2

# Scheduling and background tasks
celery==5.3.4
redis==5.0.1
apscheduler==3.10.4

# Logging
loguru==0.7.2

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Logging and monitoring
loguru==0.7.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# CORS and middleware
fastapi-cors==0.0.6

# Date and time handling
python-dateutil==2.8.2
pytz==2023.3

# JSON and data serialization
orjson==3.9.10